backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
bcrypt==4.0.1
PyJWT==2.8.0
Flask-CORS==4.0.0
reportlab==4.0.4
pandas==2.0.3
# NLP and ML libraries for advanced chatbot
spacy==3.7.2
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
nltk==3.8.1